# ddmedistock/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DDMediStock API",
    "version": "1.0.0",
    "description": "B2B medical procurement & RFQ platform API. Authentication is via an httpOnly session cookie (`ddms_session`) issued by POST /api/auth/login. State-changing requests are CSRF-protected (same-origin) and rate-limited."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin"
    }
  ],
...
  "paths": {
...
    "/api/rfqs/upload": {
      "post": {
        "tags": [
          "RFQ"
        ],
        "summary": "Upload an RFQ document (PDF/image/scanned) — OCR + AI match",
        "description": "multipart/form-data with a `file` field (and optional `title`). PDFs use the text layer when present; scanned PDFs and images are OCR'd.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "title": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "RFQ created from document",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "rfqId": {
                      "type": "string"
                    },
                    "reference": {
                      "type": "string"
                    },
                    "sourceType": {
                      "type": "string",
                      "enum": [
                        "PDF",
                        "IMAGE",
                        "CSV",
                        "EXCEL",
                        "TEXT"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "No file / invalid upload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "File too large (max 20 MB)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ddms_session"
      }
    },
...
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
    }
  }
}